feat: :sparkles: add CI workflow for build and test (#2)
* feat: :sparkles: add CI workflow for build and test

This commit introduces a new CI workflow that automates the build and test processes on push and pull request events for the main branch.

* feat: :sparkles: enhance CI workflow and add E2E testing support

Updated CI configuration to include unit and E2E tests, increased timeout, and improved build steps. Added E2E script to package.json and defined E2E task in turbo.json for better integration.

* fix: :bug: correct Playwright installation command in CI workflow

Updated the command to install Playwright browsers from 'pnpm exec' to 'npx' for better compatibility.

* fix: :bug: correct Playwright installation command in CI workflow

Replaced 'npx' with 'pnpm dlx' for installing Playwright browsers to ensure consistency with the package manager used in the workflow.

* chore: :wrench: update pnpm version to latest in CI workflow

This change ensures that the CI workflow uses the latest version of pnpm for better compatibility and features.