chore(deps): :arrow_up: upgrade dependencies (#14)
* chore(deps): :arrow_up: update husky to version 8.0.3

Upgraded husky in both package.json and pnpm-lock.yaml to ensure compatibility and access to the latest features and fixes.

* chore(deps): :arrow_up: update turbo to version 2.6.1

Upgraded turbo dependency in package.json and pnpm-lock.yaml to ensure compatibility and access to the latest features and fixes.

* chore(deps): :arrow_up: update vite to version 7.2.2

Upgraded vite in both web and shared packages to improve performance and compatibility with other dependencies.

* chore(deps): :arrow_up: update jsdom to version 27.2.0

Updated jsdom dependency in both apps/web and packages/shared to ensure compatibility with the latest features and fixes.

* docs: :books: enhance dependency upgrade instructions

Updated the steps for checking and updating dependencies to include commands for managing outdated dependencies across workspaces and specific packages.

* chore(deps): :arrow_up: upgrade dependencies

Upgraded various dependencies to the latest version.